backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg[binary]==3.3.3
python-multipart==0.0.9
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==4.2.0